refactor(sdk): extract Valkey connect preamble; ungate admin/snapshot modules (agnostic-SDK PR-6)
v0.12 backend-agnostic SDK, PR-6 (independent of PR-2/3/4/5 chain).

Shrinks `FlowFabricWorker::connect` from a ~350-line inline body into
build_client + preamble + backend-wrap. The Valkey-specific work (PING,
alive-key SET-NX, `ff:config:partitions` HGETALL, caps CSV + index
writes) moves to `crate::valkey_preamble` — byte-for-byte preserved
since scheduler reads (`ff:idx:workers` → `ff:worker:{id}:caps`)
observe the write order.

Edits:

1. New `crates/ff-sdk/src/valkey_preamble.rs` — single `run(&client,
   &config) -> PreambleOutput` entry; `#[cfg(feature =
   "valkey-default")]`-gated; the moved `read_partition_config` helper
   lives here too.

2. `FlowFabricWorker::connect` — now: build_client →
   `valkey_preamble::run` → `ValkeyBackend::from_client_and_partitions`
   → struct assembly. Public signature unchanged.

3. `lib.rs` — `pub mod admin;` + `pub mod snapshot;` ungated at module
   level. Re-exports from `admin` ungated for the transport-agnostic
   items (`FlowFabricAdminClient`, request/response shapes,
   `PartitionRotationOutcome`); `rotate_waitpoint_hmac_secret_all_partitions`
   re-export stays `valkey-default`-gated. New `pub(crate) mod
   valkey_preamble;` stays gated.

4. `admin.rs` — Option 1: keep
   `rotate_waitpoint_hmac_secret_all_partitions` (takes
   `&ferriskey::Client`) with an item-level
   `#[cfg(feature = "valkey-default")]` gate. Zero consumer-visible
   break. Snapshot.rs is already 100% trait-routed via
   `self.backend_ref().*` — no edits needed.

5. `config.rs` — new optional `WorkerConfig::partition_config:
   Option<PartitionConfig>`. `connect_with` consumes it (None →
   default); `connect` still prefers Valkey's `ff:config:partitions`
   authoritatively. Closes the v0.12.0 follow-up flagged inline in
   `connect_with` rustdoc.

6. `sqlite_only_compile_surface_tests` — new
   `admin_and_snapshot_addressable_under_sqlite_only` anchor pins
   `FlowFabricAdminClient::new` + `describe_execution` reachable under
   `--no-default-features --features sqlite`.

Validation:

- `cargo check -p ff-sdk` + `-p ff-sdk --no-default-features --features
  sqlite` — both clean.
- `cargo test --workspace --lib` (CI-shape crate set
  `-p ff-core -p ff-script -p ff-engine -p ff-scheduler -p ff-sdk
  -p ff-server --features ff-sdk/direct-valkey-claim`) — all green.
- `cargo test -p ff-sdk --no-default-features --features sqlite --lib`
  — 17 tests pass incl. the three compile-surface anchors.
- `cargo clippy` CI invocation — clean.
- `cargo tree -p ff-sdk --no-default-features --features sqlite -e
  features | grep "ff-core feature"` — only `core` (PR-0 baseline
  preserved).
- `feature-leak-guard` equivalent step — PASS (no streaming/suspension/
  budget leak into the sqlite-minimal `ff-core` edge).
- `scripts/smoke-sqlite.sh` — PASS (19s wall-clock).
- `scripts/smoke-v0.7.sh` (Valkey + Postgres) — PASS, post-`FUNCTION
  DELETE flowfabric` reload.

Non-goals: no trait method additions, no call-site ungating
(PR-3/4/5), no ferriskey removal from ff-sdk deps.
